backend/pyproject.toml
[project]
name = "notsolong-backend"
version = "0.1.0"
description = "Django REST API for the NotSoLong app"
requires-python = ">=3.13"
dependencies = [
    "django>=5.1,<6.0",
    "djangorestframework>=3.15,<4.0",
    "djangorestframework-simplejwt>=5.4,<6.0",
    "django-cors-headers>=4.4,<5.0",
    "django-environ>=0.11,<0.12",
    "inflection>=0.5,<1.0",
    "pyyaml>=6.0,<7.0",
    "uritemplate>=4.1,<5.0",
    "psycopg[binary]>=3.2,<4.0",
    "gunicorn>=23.0,<24.0",
    "whitenoise>=6.8,<7.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.6,<0.7",
    "pytest>=8.3,<9.0",
    "pytest-django>=4.8,<5.0",
    "model-bakery>=1.17,<2.0",
    "python-dotenv>=1.0,<2.0",
]

[tool.ruff]
line-length = 88
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "N", "DJ"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "notsolong.settings.dev"
python_files = ["tests.py", "test_*.py", "*_tests.py"]

[tool.setuptools]
packages = ["accounts", "api", "notsolong"]
